+++ src/components/palette/BindText.svelte
@@ -7,9 +7,14 @@
     import { locales, Projects } from '@db/Database';
     import { tick } from 'svelte';
     import Language from '@nodes/Language';
-    import { FORMATTED_SYMBOL } from '@parser/Symbols';
     import { parseFormattedLiteral } from '@parser/parseExpression';
     import { toTokens } from '@parser/toTokens';
+    import MarkupValue from '@values/MarkupValue';
+    import { FORMATTED_SYMBOL } from '@parser/Symbols';
+    import {
+        getLanguageQuoteClose,
+        getLanguageQuoteOpen,
+    } from '@locale/LanguageCode';
 
     export let property: OutputProperty;
     export let values: OutputPropertyValues;
@@ -19,6 +24,8 @@
     let project = getProject();
     let view: HTMLInputElement | undefined = undefined;
 
+    $: isMarkup = values.getValue() instanceof MarkupValue;
+
     // Whenever the text changes, update in the project.
     async function handleChange(newValue: string) {
         if ($project === undefined) return;
@@ -27,13 +34,17 @@
             $project.getBindReplacements(
                 values.getExpressions(),
                 property.getName(),
-                newValue.startsWith(FORMATTED_SYMBOL)
-                    ? parseFormattedLiteral(toTokens(newValue))
+                isMarkup
+                    ? parseFormattedLiteral(
+                          toTokens(
+                              FORMATTED_SYMBOL + newValue + FORMATTED_SYMBOL,
+                          ),
+                      )
                     : TextLiteral.make(
                           newValue,
-                          Language.make($locales.getLanguages()[0])
-                      )
-            )
+                          Language.make($locales.getLanguages()[0]),
+                      ),
+            ),
         );
 
         await tick();
@@ -41,14 +52,28 @@
     }
 </script>
 
-<TextField
-    text={values.getText()}
-    description={$locales.get((l) => l.ui.palette.field.text)}
-    placeholder={values.isEmpty()
-        ? ''
-        : $locales.getName(values.values[0].bind.names)}
-    {validator}
-    changed={handleChange}
-    bind:view
-    {editable}
-/>
+<div class="text">
+    {isMarkup
+        ? FORMATTED_SYMBOL
+        : getLanguageQuoteOpen($locales.getLocale().language)}
+    <TextField
+        text={values.getText()}
+        description={$locales.get((l) => l.ui.palette.field.text)}
+        placeholder={values.isEmpty()
+            ? ''
+            : $locales.getName(values.values[0].bind.names)}
+        {validator}
+        changed={handleChange}
+        bind:view
+        {editable}
+    />
+    {isMarkup
+        ? FORMATTED_SYMBOL
+        : getLanguageQuoteClose($locales.getLocale().language)}
+</div>
+
+<style>
+    .text {
+        display: inline;
+    }
+</style>