+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Dates are in `YYYY-MM-DD` format and versions are in [semantic versioning](http:
 -   [#310](https://github.com/wordplaydev/wordplay/issues/310): Added a new feature for detecting PII in projects and generating conflicts for them, as well as resolutions for marking text as not PII. When PII conflicts are present, projects aren't saved.
 -   [#48](https://github.com/wordplaydev/wordplay/issues/48): Added basic support for conflict resolutions.
 -   [#209](https://github.com/wordplaydev/wordplay/issues/209): Redesigned adjust buttons for literals.
+-   [#276](https://github.com/wordplaydev/wordplay/issues/276): Added weight, italic, and underline formatting controls in palette.
 -   Added explanations of save errors.
 
 ### Maintenance